garment_core/requirements/development.txt
-r base.txt
django-debug-toolbar
factory-boy
pytest
pytest-django
